chore: move dtos to common package
