+++ .husky/pre-commit
@@ -22,9 +22,6 @@ npm run generate-jest
 title "OpenAPI"
 npm run generate-openapi
 
-title "API DTOs"
-npm run generate-dtos
-
 title "Tests"
 npm run test
