+++ package.json
@@ -26,8 +26,7 @@
     "docker-rebuild": "node ./scripts/build-docker.js --nodemon",
     "env": "node ./scripts/copy-env.js",
     "env-force": "node ./scripts/copy-env.js --force",
-    "generate": "npm run generate-dtos && npm run generate-jest && npm run generate-openapi && npm run generate-references",
-    "generate-dtos": "cd packages/web && npm run copy-dtos",
+    "generate": "npm run generate-jest && npm run generate-openapi && npm run generate-references",
     "generate-jest": "node ./scripts/update-jest-config.js",
     "generate-openapi": "cd packages/api && npm run openapi",
     "generate-references": "node ./scripts/update-references.js",
