--- packages/api/src/resources/response-dto.ts
@@ -1,6 +1,0 @@
-import { PageInfoType } from '@etimo-achievements/common';
-
-export type ResponseDto<T> = {
-  data: T[];
-  pagination: PageInfoType;
-};
