+++ packages/common/src/dtos/access-token-dto.ts
