+++ packages/common/src/dtos/index.ts
@@ -1,0 +1,10 @@
+export * from './access-token-dto';
+export * from './achievement-dto';
+export * from './award-dto';
+export * from './paginated-data';
+export * from './pagination-info';
+export * from './response-dto';
+export * from './token-info-dto';
+export * from './token-validation-dto';
+export * from './user-dto';
+export * from './user-info-dto';
