+++ packages/common/src/dtos/paginated-data.ts
@@ -1,0 +1,6 @@
+import { PaginationInfo } from './pagination-info';
+
+export type PaginatedData<T> = {
+  data: T[];
+  pagination: PaginationInfo;
+};
