+++ packages/common/src/dtos/pagination-info.ts
@@ -1,0 +1,7 @@
+export type PaginationInfo = {
+  items: number;
+  totalItems: number;
+  itemsPerPage: number;
+  currentPage: number;
+  totalPages: number;
+};
