+++ packages/common/src/dtos/response-dto.ts
@@ -1,0 +1,6 @@
+import { PaginatedData } from './paginated-data';
+
+export type ResponseDto<T> = {
+  data: T[];
+  pagination: PaginatedData<T>;
+};
