+++ packages/common/src/index.ts
@@ -1,4 +1,5 @@
 export * from './context';
+export * from './dtos';
 export * from './env-helper';
 export * from './errors';
 export * from './logger';
