+++ packages/common/src/paginate.ts
@@ -1,17 +1,6 @@
-export type PageInfoType = {
-  items: number;
-  totalItems: number;
-  itemsPerPage: number;
-  currentPage: number;
-  totalPages: number;
-};
+import { PaginatedData } from '.';
 
-export type PaginationType<T> = {
-  data: T[];
-  pagination: PageInfoType;
-};
-
-export function paginate<T>(data: T[], skip: number, take: number, count: number): PaginationType<T> {
+export function paginate<T>(data: T[], skip: number, take: number, count: number): PaginatedData<T> {
   const currentPage = skip === 0 ? 1 : Math.floor(skip / take) + 1;
   return {
     data: data,
