+++ packages/service/src/modules/achievement/get-achievements-service.ts
@@ -1,4 +1,4 @@
-import { paginate, PaginationType } from '@etimo-achievements/common';
+import { paginate, PaginatedData } from '@etimo-achievements/common';
 import { AchievementRepository } from '@etimo-achievements/data';
 import { IAchievement } from '@etimo-achievements/types';
 import { ServiceOptions } from '../common/service-options';
@@ -10,7 +10,7 @@ export class GetAchievementsService {
     this.achievementRepo = options?.achievementRepository ?? new AchievementRepository();
   }
 
-  public async getMany(skip: number, take: number): Promise<PaginationType<IAchievement>> {
+  public async getMany(skip: number, take: number): Promise<PaginatedData<IAchievement>> {
     const achievements = await this.achievementRepo.getMany(skip, take);
     const count = await this.achievementRepo.count();
     return paginate(achievements, skip, take, count);
