+++ packages/service/src/modules/award/get-awards-service.ts
@@ -1,4 +1,4 @@
-import { paginate, PaginationType } from '@etimo-achievements/common';
+import { paginate, PaginatedData } from '@etimo-achievements/common';
 import { AwardRepository } from '@etimo-achievements/data';
 import { IAward } from '@etimo-achievements/types';
 import { ServiceOptions } from '../common/service-options';
@@ -10,7 +10,7 @@ export class GetAwardsService {
     this.awardRepo = options?.awardRepository ?? new AwardRepository();
   }
 
-  public async getMany(skip: number, take: number): Promise<PaginationType<IAward>> {
+  public async getMany(skip: number, take: number): Promise<PaginatedData<IAward>> {
     const users = await this.awardRepo.getAll(skip, take);
     const count = await this.awardRepo.count();
     return paginate(users, skip, take, count);
