+++ packages/service/src/modules/user/get-users-service.ts
@@ -1,4 +1,4 @@
-import { paginate, PaginationType } from '@etimo-achievements/common';
+import { paginate, PaginatedData } from '@etimo-achievements/common';
 import { UserRepository } from '@etimo-achievements/data';
 import { IUser } from '@etimo-achievements/types';
 import { ServiceOptions } from '../common/service-options';
@@ -10,7 +10,7 @@ export class GetUsersService {
     this.userRepo = options?.userRepository ?? new UserRepository();
   }
 
-  public async getAll(skip: number, take: number): Promise<PaginationType<IUser>> {
+  public async getAll(skip: number, take: number): Promise<PaginatedData<IUser>> {
     const users = await this.userRepo.getAll(skip, take);
     const count = await this.userRepo.count();
     return paginate(users, skip, take, count);
