--- packages/web/src/common/dtos/.gitattributes
@@ -1,4 +1,0 @@
-# This folder contains dtos that are copied
-# from the api project, and can hence be ignored
-# from diffing.
-* linguist-generated
