--- packages/web/src/common/dtos/award-dto.ts
@@ -1,29 +1,0 @@
-/**
- * @openapi
- * components:
- *   schemas:
- *     Award:
- *       title: Award
- *       type: object
- *       properties:
- *         id:
- *           *idProperty
- *         userId:
- *           type: string
- *           format: uuid
- *         awardedByUserId:
- *           type: string
- *           format: uuid
- *         achievementId:
- *           type: string
- *           format: uuid
- *       required:
- *         - userId
- *         - achievementId
- */
-export interface AwardDto {
-  id: string;
-  userId: string;
-  awardedByUserId: string;
-  achievementId: string;
-}
