+++ packages/web/src/features/auth/auth-slice.ts
@@ -1,7 +1,6 @@
+import { TokenInfoDto, UserInfoDto } from '@etimo-achievements/common';
 import { createSlice } from '@reduxjs/toolkit';
 import { RootState } from '../../app/reducers';
-import { TokenInfoDto } from '../../common/dtos/token-info-dto';
-import { UserInfoDto } from '../../common/dtos/user-info-dto';
 import { AuthState } from './auth-types';
 
 const initialState: AuthState = {
