+++ packages/web/src/features/auth/auth-types.ts
@@ -1,6 +1,4 @@
-import { TokenInfoDto } from '../../common/dtos/token-info-dto';
-import { UserDto } from '../../common/dtos/user-dto';
-import { UserInfoDto } from '../../common/dtos/user-info-dto';
+import { TokenInfoDto, UserDto, UserInfoDto } from '@etimo-achievements/common';
 
 export interface AuthState {
   isAuthenticated: boolean;
